requirements.txt
google-api-core==2.24.2
numpy==2.2.4
google-genai==1.10.0
openai==1.75.0
selenium==4.31.0
pillow==11.2.1
google-generativeai==0.8.4
pypdf==5.4.0